requirements.txt
streamlit>=1.28.0
requests>=2.31.0
python-dotenv>=1.0.0
Pillow>=10.0.0
streamlit-option-menu>=0.3.6
plotly>=5.15.0